Add test to compare server certificate to LDAP request
The certmonger CA dogtag-ipa-ca-renew-agent will issue a renewal
request by default when no existing certificate is found.

This is basically just a serial number to tell dogtag to issue
a new cert using the original CSR.

There have been two cases where for unknown reasons the request
didn't match the certificate so the wrong subject was being
set on renewal.