Compare serial numbers in same base, check both order for subject
certmonger stores the serial number as a hex but dogtag stores it
in ou=requests as a decimal. Convert before looking for it in the
LDAP data.

The subject order may be different between the subject in tracking
and that in the template_subject in certmonger. When comparing
test the reverse sort as well.