Add --debug and --verbose options, create failure/warning method
The debug output didn't really include a whole lot of useful
information so I added a few things.

Create a failure/warning method so that in addition to compiling
a list a debug entry will be displayed to provide context to the
failure.

validate_openssl() would blow up trying to record a failure
because it existed outside of the class.